Update Phobos to 2.072.2

+++ libphobos/libdruntime/core/sys/windows/dll.d
@@ -254,6 +254,7 @@ struct dll_aux
     }
 
     // the following structures can be found here: http://undocumented.ntinternals.net/
+    // perhaps this should be same as LDR_DATA_TABLE_ENTRY, which is introduced with PEB_LDR_DATA
     struct LDR_MODULE
     {
         LIST_ENTRY      InLoadOrderModuleList;
@@ -261,7 +262,7 @@ struct dll_aux
         LIST_ENTRY      InInitializationOrderModuleList;
         PVOID           BaseAddress;
         PVOID           EntryPoint;
-        ULONG           SizeOfImage;
+        SIZE_T          SizeOfImage;
         UNICODE_STRING  FullDllName;
         UNICODE_STRING  BaseDllName;
         ULONG           Flags;
